Mv func to delivery/util
