+++ concourse/steps/os_id.py
@@ -12,6 +12,7 @@ import ci.log
 import ci.util
 import delivery.client
 import delivery.model
+import delivery.util
 import dso.model as dm
 import github.compliance.model as gcm
 import github.compliance.issue as gciss
@@ -78,40 +79,6 @@ def base_image_os_id(
     )
 
 
-def find_branch_info(
-    os_id: um.OperatingSystemId,
-    os_infos: list[delivery.model.OsReleaseInfo],
-) -> delivery.model.OsReleaseInfo:
-    if not os_id.ID:
-        return None # os-id could not be determined
-
-    os_version = os_id.VERSION_ID
-
-    def version_candidates():
-        yield os_version
-        yield f'v{os_version}'
-
-        parts = os_version.split('.')
-
-        if len(parts) == 1:
-            return
-
-        yield parts[0]
-        yield 'v' + parts[0]
-
-        yield '.'.join(parts[:2]) # strip parts after minor
-        yield 'v' + '.'.join(parts[:2]) # strip parts after minor
-
-    candidates = tuple(version_candidates())
-
-    for os_info in os_infos:
-        for candidate in candidates:
-            if os_info.name == candidate:
-                return os_info
-
-    logger.warning(f'did not find branch-info for {os_id=}')
-
-
 def scan_result_group_collection_for_outdated_os_ids(
     results: tuple[gcm.OsIdScanResult],
     delivery_svc_client: delivery.client.DeliveryServiceClient,
@@ -124,7 +91,7 @@ def scan_result_group_collection_for_outdated_os_ids(
     }
 
     def branch_reached_eol(os_id: um.OperatingSystemId):
-        branch_info = find_branch_info(
+        branch_info = delivery.util.find_branch_info(
             os_id=os_id,
             os_infos=os_infos[os_id.ID],
         )
@@ -134,7 +101,7 @@ def scan_result_group_collection_for_outdated_os_ids(
         return branch_info.reached_eol()
 
     def update_available(os_id: um.OperatingSystemId):
-        branch_info = find_branch_info(
+        branch_info = delivery.util.find_branch_info(
             os_id=os_id,
             os_infos=os_infos[os_id.ID],
         )
